ExampleApp: Move DomainHandlers to Application layer
